move parsing to io.rs
